Separate tests by eventType
